Use explicit integer sizes in memory

+++ src/memory.c
@@ -15,15 +15,15 @@ void do_dma(void)
     mega65_io_enable();
 
     //  for(i=0;i<24;i++)
-    //    printf("%02x ",PEEK(i+(unsigned int)&dmalist));
+    //    printf("%02x ",PEEK(i+(uint16_t)&dmalist));
     //  printf("\n");
     //  while(1) continue;
 
     // Now run DMA job (to and from anywhere, and list is in low 1MB)
     POKE(0xd702U, 0);
     POKE(0xd704U, 0x00); // List is in $00xxxxx
-    POKE(0xd701U, ((unsigned int)&dmalist) >> 8);
-    POKE(0xd705U, ((unsigned int)&dmalist) & 0xff); // triggers enhanced DMA
+    POKE(0xd701U, ((uint16_t)&dmalist) >> 8);
+    POKE(0xd705U, ((uint16_t)&dmalist) & 0xff); // triggers enhanced DMA
 }
 
 uint8_t dma_peek(uint32_t address)
@@ -47,7 +47,7 @@ uint8_t dma_peek(uint32_t address)
     dmalist.count = 1;
     dmalist.source_addr = address & 0xffff;
     dmalist.source_bank = (address >> 16) & 0x0f;
-    dmalist.dest_addr = (unsigned int)&dma_byte;
+    dmalist.dest_addr = (uint16_t)&dma_byte;
     dmalist.dest_bank = 0;
 
     do_dma();
@@ -96,7 +96,7 @@ void dma_poke(uint32_t address, uint8_t value)
     dma_byte = value;
     dmalist.command = 0x00; // copy
     dmalist.count = 1;
-    dmalist.source_addr = (unsigned int)&dma_byte;
+    dmalist.source_addr = (uint16_t)&dma_byte;
     dmalist.source_bank = 0;
     dmalist.dest_addr = address & 0xffff;
     dmalist.dest_bank = (address >> 16) & 0x0f;
@@ -106,7 +106,7 @@ void dma_poke(uint32_t address, uint8_t value)
 }
 
 void lcopy(
-    uint32_t source_address, uint32_t destination_address, unsigned int count)
+    uint32_t source_address, uint32_t destination_address, uint16_t count)
 {
     dmalist.option_0b = 0x0b;
     dmalist.option_80 = 0x80;
@@ -138,7 +138,7 @@ void lcopy(
     return;
 }
 
-void lfill(uint32_t destination_address, uint8_t value, unsigned int count)
+void lfill(uint32_t destination_address, uint8_t value, uint16_t count)
 {
     dmalist.option_0b = 0x0b;
     dmalist.option_80 = 0x80;
@@ -164,7 +164,7 @@ void lfill(uint32_t destination_address, uint8_t value, unsigned int count)
     return;
 }
 
-void lfill_skip(uint32_t destination_address, uint8_t value, unsigned int count,
+void lfill_skip(uint32_t destination_address, uint8_t value, uint16_t count,
     uint8_t skip)
 {
     dmalist.option_0b = 0x0b;
